# Remove permute() and alternate method of getting g() inlined leads to better performance