--- a/gen/gen.go
+++ b/gen/gen.go
@@ -24,16 +24,16 @@ func main() {
 		fmt.Printf(`// round%d
 		
 	// Mix the columns.
-	s[0], s[4], s[8], s[12] = g(s[0], s[4], s[8], s[12], n.block[%d], n.block[%d])
-	s[1], s[5], s[9], s[13] = g(s[1], s[5], s[9], s[13], n.block[%d], n.block[%d])
-	s[2], s[6], s[10], s[14] = g(s[2], s[6], s[10], s[14], n.block[%d], n.block[%d])
-	s[3], s[7], s[11], s[15] = g(s[3], s[7], s[11], s[15], n.block[%d], n.block[%d])
+	s0, s4, s8, s12 = g(s0, s4, s8, s12, n.block[%d], n.block[%d])
+	s1, s5, s9, s13 = g(s1, s5, s9, s13, n.block[%d], n.block[%d])
+	s2, s6, s10, s14 = g(s2, s6, s10, s14, n.block[%d], n.block[%d])
+	s3, s7, s11, s15 = g(s3, s7, s11, s15, n.block[%d], n.block[%d])
 
 	// Mix the diagonals.
-	s[0], s[5], s[10], s[15] = g(s[0], s[5], s[10], s[15], n.block[%d], n.block[%d])
-	s[1], s[6], s[11], s[12] = g(s[1], s[6], s[11], s[12], n.block[%d], n.block[%d])
-	s[2], s[7], s[8], s[13] = g(s[2], s[7], s[8], s[13], n.block[%d], n.block[%d])
-	s[3], s[4], s[9], s[14] = g(s[3], s[4], s[9], s[14], n.block[%d], n.block[%d])
+	s0, s5, s10, s15 = g(s0, s5, s10, s15, n.block[%d], n.block[%d])
+	s1, s6, s11, s12 = g(s1, s6, s11, s12, n.block[%d], n.block[%d])
+	s2, s7, s8, s13 = g(s2, s7, s8, s13, n.block[%d], n.block[%d])
+	s3, s4, s9, s14 = g(s3, s4, s9, s14, n.block[%d], n.block[%d])
 `, x,
 			m[0], m[1], m[2], m[3], m[4], m[5], m[6], m[7], m[8], m[9], m[10], m[11], m[12], m[13], m[14], m[15])
 		permute(&m)