* Factors out the need for permute(), by generating code for each round.
 * g() inlines when pass the values instead of an array pointer and indices. This allows putting state across 16 variables instead of [16]uint32 which then causes the compiler to use more registers.

```
name      old time/op    new time/op    delta
Write-4     6.18ns ± 0%    3.41ns ± 0%  -44.93%  (p=0.000 n=9+10)
Sum256-4    6.21µs ± 0%    3.51µs ± 0%  -43.43%  (p=0.000 n=9+9)
XOF-4       5.73ns ± 0%    3.13ns ± 0%  -45.36%  (p=0.000 n=9+10)

name      old speed      new speed      delta
Write-4    162MB/s ± 0%   294MB/s ± 0%  +81.53%  (p=0.000 n=9+10)
XOF-4      174MB/s ± 0%   319MB/s ± 0%  +83.04%  (p=0.000 n=9+10)

name      old alloc/op   new alloc/op   delta
Write-4      0.00B          0.00B          ~     (all equal)
Sum256-4     0.00B          0.00B          ~     (all equal)
XOF-4        0.00B          0.00B          ~     (all equal)

name      old allocs/op  new allocs/op  delta
Write-4       0.00           0.00          ~     (all equal)
Sum256-4      0.00           0.00          ~     (all equal)
XOF-4         0.00           0.00          ~     (all equal)
```