@lukechampine 2020-01-13
Thanks again -- another huge increase!

I decided not to keep the codegen, since it's not likely to ever change, and the generated code isn't terribly hard to verify. However, if we someone wants to add an asm implementation, I would push for codegen there -- a tool like [avo](https://github.com/mmcloughlin/avo) would make the asm much easier to to review and verify.